+++ src/main/java/chess/controller/ChessGameController.java
@@ -4,10 +4,10 @@ import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.ScoreCalculator;
 import chess.domain.board.ChessBoard;
-import chess.domain.board.DBRepository;
-import chess.domain.dbUtils.BoardDao;
-import chess.domain.dbUtils.DBConnectionUtils;
-import chess.domain.dbUtils.GameDao;
+import chess.db.DBRepository;
+import chess.db.BoardDao;
+import chess.db.DBConnectionUtils;
+import chess.db.GameDao;
 import chess.domain.position.Position;
 import chess.dto.PositionParser;
 import chess.dto.CommandDto;
