+++ src/main/java/chess/db/GameDao.java
@@ -1,4 +1,4 @@
-package chess.domain.dbUtils;
+package chess.db;
 
 import chess.domain.Color;
 import java.sql.Connection;
